Complete conversions

+++ exercises/23_conversions/using_as.rs
@@ -4,8 +4,9 @@
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    // TODO: Make a conversion before dividing.
-    total / values.len()
+    // DONE: Make a conversion before dividing.
+
+    total / values.len() as f64
 }
 
 fn main() {
